+++ tests/text_to_ast/desktop_set.rs
@@ -110,8 +110,7 @@ fn hashtag() {
 #[test]
 fn german_umlaut_hashtag() {
     let input = "#bücher #Ängste";
-    // revert this back to assert_eq, once implemented see https://github.com/deltachat/message-parser/issues/8 for more info
-    assert_ne!(
+    assert_eq!(
         parse_desktop_set(input),
         vec![Tag("#bücher"), Text(" "), Tag("#Ängste")]
     );